move test_worker2 to test_worker
